requirements.txt
ciscoconfparse==1.4.11    
netmiko==3.1.0
PTable==0.9.2
PyYAML==5.3
textfsm==1.1.0
